Corrections pass 1
